moving prettierVersionSatisfies as an utility function
